requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==3.0.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
fastapi==0.112.1
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
jiter==0.5.0
multidict==6.0.5
oauthlib==3.2.2
openai==1.41.0
packaging==24.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.27.3
py==1.11.0
py-cord==2.6.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pymongo==4.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
pytest==8.3.2
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
yarl==1.9.4
sse-starlette==2.1.3
langchain==0.2.14
langgraph==0.2.5
langserve==0.2.2
langsmith==0.1.100
langchain-openai==0.1.22
langchain-anthropic==0.1.23
opuslib==3.0.1
loguru==0.7.2
pytz==2024.1
pydub==0.25.1
